Add new step for selecting AMC Instances.
