+++ source/website/src/store/mutations.js
@@ -14,4 +14,7 @@ export default {
   updateS3key (state, value) {
     state.s3key = value
   },
+  updateDestinations (state, value) {
+    state.destinations = value
+  }
 }
