+++ source/website/src/views/Step5.vue
@@ -7,209 +7,84 @@ SPDX-License-Identifier: Apache-2.0
   <div>
     <div class="headerTextBackground">
       <Header />
+      <b-modal
+        v-model="showModal"
+        :title="modal_title"
+        ok-only
+        @ok="hideModal"
+      >
+        {{ response }}
+      </b-modal>
       <b-container fluid>
         <b-row style="text-align: left">
           <b-col cols="2">
             <Sidebar :is-step5-active="true" />
           </b-col>
           <b-col cols="10">
+            <h3>Confirm Details</h3>
             <b-row>
-              <b-col>
-                <h3>Phase 1: Datasets defined</h3>
+              <b-col sm="7">
+                Click Submit to record this dataset in AMC.
               </b-col>
-              <b-col align="right">
-                <b-button @click="list_datasets()">
-                  Refresh
-                </b-button>
+              <b-col sm="3" align="right">
+                <button type="submit" class="btn btn-outline-primary mb-2" @click="$router.push('Step3')">
+                  Previous
+                </button> &nbsp;
+                <button type="submit" class="btn btn-primary mb-2" @click="onSubmit">
+                  Submit
+                  <b-spinner v-if="isBusy" style="vertical-align: sub" small label="Spinning"></b-spinner>
+                </button>
               </b-col>
             </b-row>
-            <b-table 
-              ref="datasetTable"
-              select-mode="single"
-              selectable
-              responsive="sm"
-              small
-              :fields="dataset_fields"
-              :items="datasets"
-              :busy="isBusy"
-              :per-page="perPagePhase1"
-              :current-page="currentPagePhase1"
-              sort-by="updatedTime"
-              :sort-desc="true"
-              show-empty
-              @row-selected="onRowSelected"
-            >
-              <template #empty="scope">
-                {{ scope.emptyText }}
-              </template>
-              <template #cell(createdtime)="data">
-                {{ new Date(data.item.createdTime).toLocaleString() }}
-              </template>
-              <template #cell(updatedtime)="data">
-                {{ new Date(data.item.updatedTime).toLocaleString() }}
-              </template>
-              <template #cell(Actions)="data">
-                <b-link
-                  class="text-danger"
-                  @click="deleteDataset(`${data.item.dataSetId}`)"
-                >
-                  Delete
-                </b-link>
-              </template>
-              <template #cell(selected)="{ rowSelected }">
-                <template v-if="rowSelected">
-                  <span aria-hidden="true">&check;</span>
-                  <span class="sr-only">Selected</span>
-                </template>
-                <template v-else>
-                  <span aria-hidden="true">&nbsp;</span>
-                  <span class="sr-only">Not selected</span>
-                </template>
-              </template>
-              <template #table-busy>
-                <div class="text-center my-2">
-                  <b-spinner class="align-middle"></b-spinner>
-                  <strong>&nbsp;&nbsp;Loading...</strong>
-                </div>
-              </template>
-            </b-table>
-            <b-pagination
-              v-if="datasets.length > perPagePhase1"
-              v-model="currentPagePhase1"
-              align="center"
-              :per-page="perPagePhase1"
-              :total-rows="rowsPhase1"
-              aria-controls="shotTable"
-            ></b-pagination>
-            <br>
             <b-row>
-              <b-col>
-                <h3>Phase 2: Datasets transformed</h3>
-              </b-col>
-              <b-col align="right">
-                <b-button @click="get_etl_jobs()">
-                  Refresh
-                </b-button>
-              </b-col>
-            </b-row>
-            <b-table
-              small
-              responsive="sm"
-              :fields="etl_fields"
-              :items="etl_jobs"
-              :busy="isBusy2"
-              :per-page="perPagePhase2"
-              :current-page="currentPagePhase2"
-              sort-by="StartedOn"
-              :sort-desc="true"
-              show-empty
-            >
-              <template #cell(id)="data">
-                ...{{ data.item.Id.substr(-8) }}
-              </template>
-              <template #cell(JobRunState)="data">
-                {{ data.item.JobRunState }}
-              </template>
-              <template #cell(show_details)="row">
-                <b-form-checkbox @change="row.toggleDetails">
-                </b-form-checkbox>
-              </template>
-              <template #row-details="data">
-                <b-card>
-                  <b-row v-for="(item, key) in data.item" :key="key" class="mb-2">
-                    <b-col sm="3" class="text-sm-right">
-                      <strong>{{ key }}:</strong>
-                    </b-col>
-                    <b-col>
-                      {{ item }}
-                    </b-col>
-                  </b-row>
-                </b-card>
-              </template>
-              <template #empty>
-                No ETL jobs have started yet.
-              </template>
-              <template #table-busy>
-                <div class="text-center my-2">
-                  <b-spinner class="align-middle"></b-spinner>
-                  <strong>&nbsp;&nbsp;Loading...</strong>
-                </div>
-              </template>
-            </b-table>
-            <b-pagination 
-              v-if="etl_jobs.length > perPagePhase2"
-              v-model="currentPagePhase2"
-              align="center"
-              :per-page="perPagePhase2"
-              :total-rows="rowsPhase2"
-              aria-controls="shotTable"
-            ></b-pagination>
-            <br>
-            <b-row>
-              <b-col>
-                <h3>Phase 3: Datasets uploaded</h3>
-                <div v-if="selected_dataset">
-                  Showing uploads for {{ selected_dataset }}
-                </div>
-                <div v-else>
-                  Please select a dataset from the table in Phase 1.
-                </div>
-              </b-col>
-              <b-col v-if="selected_dataset" align="right">
+              <b-col cols="7">
+                <h5>Input files:</h5>
+                <ul>
+                  <li v-for="item in s3key.split(',')">
+                    {{ "s3://" + DATA_BUCKET_NAME + "/" + item }}
+                  </li>
+                </ul>
+                <h5>Destinations:</h5>
+                <ul>
+                <li v-for="amc_instance in destinations">
+                  {{ amc_instance }}
+                </li>
+                </ul>
                 <br>
-                <b-button @click="listDatasetUploads(`${selected_dataset}`)">
-                  Refresh
-                </b-button>
+                <h5>Dataset Attributes:</h5>
+                <b-table
+                  small
+                  outlined
+                  :items="dataset.other_attributes"
+                  thead-class="hidden_header"
+                  show-empty
+                >
+                </b-table>
+                <template #empty="scope">
+                  {{ scope.emptyText }}
+                </template>
               </b-col>
             </b-row>
-            <div v-if="selected_dataset">
-              <b-table
-                :items="uploads"
-                :fields="upload_fields"
-                :busy="isBusy3"
-                :per-page="perPagePhase3"
-                :current-page="currentPagePhase3"
-                sort-by="dateCreated"
-                :sort-desc="true"
-                show-empty
-                small
-                responsive="sm"
-              >
-                <template #empty>
-                  The dataset {{ selected_dataset }} has not been uploaded.
-                </template>
-                <template #table-busy>
-                  <div class="text-center my-2">
-                    <b-spinner class="align-middle"></b-spinner>
-                    <strong>&nbsp;&nbsp;Loading...</strong>
-                  </div>
-                </template>
-                <template #cell(show_details)="row">
-                  <b-form-checkbox @change="row.toggleDetails">
-                  </b-form-checkbox>
-                </template>
-                <template #row-details="data">
-                  <b-card>
-                    <b-row v-for="(item, key) in data.item" :key="key" class="mb-2">
-                      <b-col sm="3" class="text-sm-right">
-                        <strong>{{ key }}:</strong>
-                      </b-col>
-                      <b-col>
-                        {{ item }}
-                      </b-col>
-                    </b-row>
-                  </b-card>
-                </template>
-              </b-table>
-              <b-pagination 
-                v-if="uploads.length > perPagePhase3"
-                v-model="currentPagePhase3"
-                align="center"
-                :per-page="perPagePhase3"
-                :total-rows="rowsPhase3"
-                aria-controls="shotTable"
-              ></b-pagination>
-            </div>
+            <b-row>
+              <b-col cols="10">
+                <h5>Columns:</h5>
+                <b-table 
+                  v-if="dataset.columns && dataset.columns.length > 0"
+                  small
+                  outlined
+                  :fields="column_fields"
+                  :items="dataset.columns"
+                >
+                </b-table>
+                <b-table
+                  v-else
+                  small
+                  outlined
+                  :items="dataset.columns"
+                  show-empty
+                ></b-table>
+              </b-col>
+            </b-row>
           </b-col>
         </b-row>
       </b-container>
@@ -220,6 +95,7 @@ SPDX-License-Identifier: Apache-2.0
 <script>
   import Header from '@/components/Header.vue'
   import Sidebar from '@/components/Sidebar.vue'
+  import {mapState} from "vuex";
 
   export default {
     name: "Step5",
@@ -228,56 +104,42 @@ SPDX-License-Identifier: Apache-2.0
     },
     data() {
       return {
-        currentPagePhase3: 1,
-        perPagePhase3: 10,
-        currentPagePhase2: 1,
-        perPagePhase2: 5,
-        currentPagePhase1: 1,
-        perPagePhase1: 10,
-        datasets: [],
-        selected_dataset: '',
-        dataset_fields: [
-          {key: 'selected'},
-          {key: 'dataSetId', sortable:true},
-          {key: 'description'},
-          {key: 'dataSetType', sortable:true},
-          {key: 'fileFormat', sortable:true},
-          {key: 'createdTime', sortable:true},
-          {key: 'updatedTime', sortable:true},
-          {key: 'Actions'}
-        ],
-        etl_jobs: [],
-        etl_fields: ['DatasetId', 'Id', 'StartedOn', 'CompletedOn', 'ExecutionTime', 'JobRunState', 'show_details'],
-        uploads: [],
-        upload_fields: [
-          {key: "dateCreated", label: "Date Created", sortable: true},
-          {key: "totalFileCount", label: "Total Files"},
-          {key: "errorFileCount", label: "Bad Files"},
-          {key: "rowsAcceptedTotal", label: "Rows Accepted"},
-          {key: "rowsDroppedTotal", label: "Rows Dropped"},
-          {key: "rowsWithResolvedIdentity", label: "Identities Resolved"},
-          {key: "sourceFileS3Key", label: "Source File", sortable: true},
-          {key: "uploadId", label: "Upload Id"},
-          {key: "status", label: "Status"},
-          {key: "show_details", label: "Show Details"}
-        ],
+        column_fields: ['name', 'description', 'dataType', 'columnType', 'nullable', 'isMainUserId', 'isMainUserIdType', 'isMainUserId', 'externalUserIdType.identifierType','isMainEventTime'],
+        dataset_fields: [{key: '0', label: 'Name'}, {key: '1', label: 'Value'}],
         isBusy: false,
-        isBusy2: false,
-        isBusy3: false,
+        showModal: false,
+        modal_title: '',
         isStep5Active: true,
         response: ''
       }
     },
     computed: {
-      rowsPhase1() {
-        return this.datasets.length
+      ...mapState(['deleted_columns', 'dataset_definition', 's3key', 'destinations']),
+      encryption_key() {
+        if (this.CUSTOMER_MANAGED_KEY === "") {
+          return "default"
+        } else {
+          return this.CUSTOMER_MANAGED_KEY
+        }
       },
-      rowsPhase2() {
-        return this.etl_jobs.length
+      pii_fields() {
+         return (this.dataset.columns.filter(x => x.externalUserIdType).map(x => (new Object( {'column_name':x.name, 'pii_type': x.externalUserIdType.identifierType}))))
       },
-      rowsPhase3() {
-        return this.uploads.length
+      timestamp_column_name() {
+        const timestamp_column = this.dataset.columns.filter(x => x.isMainEventTime).map(x => x.name)
+        // The Glue ETL job requires timestamp_column_name to be an empty string 
+        // for all DIMENSION datasets.
+        const dataset_type = this.dataset_definition['dataSetType']
+        if (dataset_type == 'FACT' && timestamp_column && timestamp_column.length > 0)
+          return timestamp_column[0]
+        else
+          return ''
       },
+      dataset() {
+        let {columns, ...other_attributes} = this.dataset_definition
+        other_attributes['encryption_mode'] = this.ENCRYPTION_MODE
+        return {"columns": columns, "other_attributes": Object.entries(other_attributes)}
+      }
     },
     deactivated: function () {
       console.log('deactivated');
@@ -287,117 +149,68 @@ SPDX-License-Identifier: Apache-2.0
     },
     created: function () {
       console.log('created')
-      this.list_datasets()
-      this.get_etl_jobs()
     },
     methods: {
-      onRowSelected(items) {
-        if (items.length > 0) {
-          this.selected_dataset = items[0].dataSetId
-          this.listDatasetUploads(this.selected_dataset)
-        } else {
-          this.uploads = []
-          this.selected_dataset=''
-        }
+      hideModal() {
+        this.showModal = false
       },
-      async deleteDataset(dataSetId) {
-        this.datasets = this.datasets.filter(x => x.dataSetId !== dataSetId)
-        await this.delete_dataset({'dataSetId': dataSetId})
+      onSubmit() {
+        this.send_request('POST', 'create_dataset', {'body': this.dataset_definition})
       },
-      async delete_dataset(data) {
+      async send_request(method, resource, data) {
+        console.log("sending " + method + " " + resource + " " + JSON.stringify(data))
         const apiName = 'amcufa-api'
         let response = ""
-        const method = 'POST'
-        const resource = 'delete_dataset'
-        try {
-          console.log("sending " + method + " " + resource + " " + JSON.stringify(data))
-          let requestOpts = {
-            headers: {'Content-Type': 'application/json'},
-            body: data
-          };
-          response = await this.$Amplify.API.post(apiName, resource, requestOpts);
-          console.log(response)
-        }
-        catch (e) {
-          console.log("ERROR: " + e.response.data.message)
-          this.response = e.response.data.message
-        }
-      },
-      async listDatasetUploads(dataSetId) {
-        this.selected_dataset = dataSetId
-        await this.list_uploads({'dataSetId': dataSetId})
-      },
-      async list_uploads(data) {
-        this.uploads = []
-        const apiName = 'amcufa-api'
-        const method = 'POST'
-        const resource = 'list_uploads'
-        this.isBusy3 = true;
-        try {
-          console.log("sending " + method + " " + resource + " " + JSON.stringify(data))
-          let requestOpts = {
-            headers: {'Content-Type': 'application/json'},
-            body: data  
-          };
-          do {
-            const response = await this.$Amplify.API.post(apiName, resource, requestOpts);
-            this.uploads.push(...response.uploads);
-            console.log(this.uploads)
-            data.nextToken = response.nextToken;
-            requestOpts.body = data;
-          } while (data.nextToken);
-        }
-        catch (e) {
-          console.log("ERROR: " + e.response.data.message)
-          this.isBusy3 = false;
-          this.response = e.response.data.message
-        }
-        this.isBusy3 = false;
-      },
-      async list_datasets() {
-        const apiName = 'amcufa-api'
-        let response = ""
-        const method = 'GET'
-        const resource = 'list_datasets'
         this.isBusy = true;
         try {
           if (method === "GET") {
-            console.log("sending " + method + " " + resource)
             response = await this.$Amplify.API.get(apiName, resource);
+          } else if (method === "POST") {
+            let requestOpts = {
+              headers: {'Content-Type': 'application/json'},
+              body: data
+            };
+            response = await this.$Amplify.API.post(apiName, resource, requestOpts);
+          }
+          console.log(JSON.stringify(response))
+          console.log("Dataset defined successfully")
+          
+          // Start Glue ETL job now that the dataset has been accepted by AMC
+          let s3keysList = this.s3key.split(',').map((item) => item.trim())
+
+          for (let key of s3keysList) {
+            console.log("Starting Glue ETL job for s3://" + this.DATA_BUCKET_NAME + "/" + key)
+            resource = 'start_amc_transformation'
+            data = {'sourceBucket': this.DATA_BUCKET_NAME, 'sourceKey': key, 'outputBucket': this.ARTIFACT_BUCKET_NAME, 'piiFields': JSON.stringify(this.pii_fields),'deletedFields': JSON.stringify(this.deleted_columns), 'timestampColumn': this.timestamp_column_name, 'datasetId': this.dataset_definition.dataSetId, 'period': this.dataset_definition.period}
+            let requestOpts = {
+              headers: {'Content-Type': 'application/json'},
+              body: data
+            };
+            console.log("POST " + resource + " " + JSON.stringify(requestOpts))
+            response = await this.$Amplify.API.post(apiName, resource, requestOpts);
             console.log(response)
-            this.datasets = response.dataSets
-          } 
+            console.log(JSON.stringify(response))
+            console.log("Started Glue ETL job")
+          }
+          
+          // Navigate to next step
+          this.$router.push('Step5')
         }
         catch (e) {
-          console.log("ERROR: " + e.response.data.message)
+          this.modal_title = e.response.status + " " + e.response.statusText
+          console.log("ERROR: " + this.modal_title)
           this.isBusy = false;
-          this.response = e.response.data.message
+          this.response = JSON.stringify(e.response.data)  
+          this.showModal = true
         }
         this.isBusy = false;
-      },
-      async get_etl_jobs() {
-        this.isBusy2 = true;
-        this.etl_jobs = []
-        const apiName = 'amcufa-api'
-        let response = ""
-        const method = 'GET'
-        const resource = 'get_etl_jobs'
-        try {
-          if (method === "GET") {
-            console.log("sending " + method + " " + resource)
-            response = await this.$Amplify.API.get(apiName, resource);
-            console.log(response)
-            this.etl_jobs = response.JobRuns
-          } 
-        }
-        catch (e) {
-          console.log("ERROR: " + e.response.data.message)
-          this.isBusy2 = false;
-          this.response = e.response.data.message
-        }
-        this.isBusy2 = false;
       }
     }
   }
 </script>
 
+<style>
+.hidden_header {
+  display: none;
+}
+</style>
